added authentication and authorization

+++ app/controllers/workouts_controller.rb
@@ -11,8 +11,12 @@ class WorkoutsController < ApplicationController
   end
 
   def destroy
-    Workout.find(params[:id]).exercise.destroy
-    redirect_to current_user.current_plan, notice: 'Successfully deleted your workout.'
+    if current_user == Workout.find(params[:id]).exercise.workout_plan.user
+      Workout.find(params[:id]).exercise.destroy
+      redirect_to current_user.current_plan, notice: 'Successfully deleted your workout.'
+    else
+      redirect_to root_path, alert: 'Access Denied'
+    end
   end
 
   private
